chore(xtest): New checkout layout
To support running multiple versions of the same repo,
we now check out source code like this:

```
/xtest/sdk/[lang]/src/[version]/[source code at version]
```

The new `checkout-all` script will do this for the `main` branch.

To build all checked out sdks, use the Makefile in `xtest/sdk`.
This creates a `dist/[version]` subfolder for each checked out version.

This also removes npm xtest support; this distracts from pytest, which is needed, by making it appear that everything is running in node, which it isn't